+++ app/models/parser/base.rb
@@ -157,8 +157,7 @@ module Parser
 
     def twitter_author_url(username)
       return if bad_username?(username)
-      # username[1..]: remove the @ from the username
-        "https://twitter.com/" + username[1..]
+        "https://twitter.com/" + username.gsub("@","")
     end
 
     def bad_username?(value)
