+++ app/models/parser/twitter_item.rb
@@ -18,50 +18,42 @@ module Parser
 
     # Main function for class
     def parse_data_for_parser(_doc, _original_url, _jsonld_array)
-      @url.gsub!(/(%23|#)!\//, '')
-      @url.gsub!(/\s/, '')
-      @url = replace_subdomain_pattern(url)
-      
-      parts = url.match(TWITTER_ITEM_URL)
-      user, id = parts['user'], parts['id']
-      
-      @parsed_data['raw']['api'] = tweet_lookup(id)
-
-      @parsed_data[:error] = parsed_data['raw']['api']['errors']
-
-      if @parsed_data[:error] 
-        title = url
-        description = ''
-        author_picture = ''
-        published_at = ''
-        html = ''
-        author_name = user
-        author_url = get_author_url(user)
-      elsif @parsed_data[:error].nil?
-        raw_data = parsed_data['raw']['api']['data'][0]
-        raw_user_data = parsed_data['raw']['api']['includes']['users'][0]
-
-        title = raw_data['text'].squish
-        description = title
-        author_picture = raw_user_data['profile_image_url'].gsub('_normal', '')
-        published_at = raw_data['created_at']
-        html = html_for_twitter_item(url)
-        author_name = raw_user_data['name']
-        author_url = get_author_url(user) || raw_user_data['url'] || RequestHelper.top_url(url)
+      handle_exceptions(StandardError) do
+        @url.gsub!(/(%23|#)!\//, '')
+        @url.gsub!(/\s/, '')
+        @url = replace_subdomain_pattern(url)
+        
+        parts = url.match(TWITTER_ITEM_URL)
+        user, id = parts['user'], parts['id']
+        
+        @parsed_data.merge!(          
+          external_id: id,
+          username: '@' + user,
+          author_url: get_author_url(user)
+        )
+          
+        @parsed_data['raw']['api'] = tweet_lookup(id)
+        @parsed_data[:error] = parsed_data.dig('raw', 'api', 'errors')
+        
+        if @parsed_data[:error] 
+          @parsed_data.merge!(
+            author_name: user,
+          )
+        elsif @parsed_data[:error].nil?
+          raw_data = parsed_data.dig('raw','api','data',0)
+          raw_user_data = parsed_data.dig('raw','api','includes','users',0)
+          
+          @parsed_data.merge!({
+            picture: get_twitter_item_picture(parsed_data),
+            title: raw_data['text'].squish,
+            description: raw_data['text'].squish,
+            author_picture: raw_user_data['profile_image_url'].gsub('_normal', ''),
+            published_at: raw_data['created_at'],
+            html: html_for_twitter_item(url),
+            author_name: raw_user_data['name'],
+          })  
+        end
       end
- 
-      @parsed_data.merge!({
-        external_id: id,
-        username: '@' + user,
-        title: title,
-        description: description,
-        picture: get_twitter_item_picture(parsed_data),
-        author_picture: author_picture,
-        published_at: published_at,
-        html: html,
-        author_name: author_name,
-        author_url: author_url
-      })
       parsed_data
     end
 
@@ -70,9 +62,9 @@ module Parser
     end
 
     def get_twitter_item_picture(parsed_data)
-      return unless parsed_data['raw']['api']['includes']
-      media = parsed_data['raw']['api']['includes']['media']
-      media.nil? ? nil : media[0]['url']
+      return unless parsed_data.dig('raw', 'api', 'includes')
+      item_media = parsed_data.dig('raw', 'api', 'includes', 'media')
+      item_media.dig(0, 'url') if item_media
     end
 
     def html_for_twitter_item(url)
